Break up the types
