+++ types.go
@@ -1,90 +1,8 @@
 package wellnessliving
 
 import (
-	"encoding/json"
 	"fmt"
-	"strconv"
-	"time"
 )
-
-type StringMap map[string]string
-
-type Date struct {
-	time.Time
-}
-
-func (d *Date) UnmarshalJSON(contents []byte) error {
-	var v string
-	err := json.Unmarshal(contents, &v)
-	if err != nil {
-		return err
-	}
-
-	location, err := time.LoadLocation("GMT")
-	if err != nil {
-		return err
-	}
-
-	d.Time, err = time.ParseInLocation("2006-01-02", v, location)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-type DateTime struct {
-	time.Time
-}
-
-func (d *DateTime) UnmarshalJSON(contents []byte) error {
-	var v string
-	err := json.Unmarshal(contents, &v)
-	if err != nil {
-		return err
-	}
-
-	location, err := time.LoadLocation("GMT")
-	if err != nil {
-		return err
-	}
-
-	d.Time, err = time.ParseInLocation("2006-01-02 15:04:05", v, location)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-type Currency float64
-
-func (d *Currency) UnmarshalJSON(contents []byte) error {
-	var v string
-	err := json.Unmarshal(contents, &v)
-	if err != nil {
-		return err
-	}
-
-	f, err := strconv.ParseFloat(v, 64)
-	if err != nil {
-		return err
-	}
-	*d = Currency(f)
-	return nil
-}
-
-func (m *StringMap) UnmarshalJSON(contents []byte) error {
-	if string(contents) == "[]" {
-		*m = map[string]string{}
-	} else {
-		v := map[string]string{}
-		err := json.Unmarshal(contents, &v)
-		if err != nil {
-			return err
-		}
-		*m = v
-	}
-	return nil
-}
 
 type BaseResponse struct {
 	Status  string `json:"status"`
